fix case

+++ springside3/trunk/examples/mini-service/src/main/java/org/springside/examples/miniservice/service/user/UserManager.java
@@ -49,6 +49,6 @@ public class UserManager {
 		if (StringUtils.isBlank(loginName) || StringUtils.isBlank(password))
 			return false;
 
-		return ((Integer) userDao.findUnique(UserDao.QUERY_BY_LNAME_PASSWD, loginName, password) == 1);
+		return ((Long) userDao.findUnique(UserDao.QUERY_BY_LNAME_PASSWD, loginName, password) == 1);
 	}
 }
